refactor to batch
All other adapters use a batch class to do the actual work with their
schedulers, so this is a refactor to use that same pattern.
