+++ lib/ood_core/job/adapters/kubernetes.rb
@@ -1,48 +1,32 @@
+require "ood_core/refinements/hash_extensions"
+require "ood_core/refinements/array_extensions"
+
 module OodCore
   module Job
     class Factory
       using Refinements::HashExtensions
 
       def self.build_kubernetes(config)
-        Adapters::Kubernetes.new(config)
+        batch = Adapters::Kubernetes::Batch.new(config.to_h.symbolize_keys, Adapters::Kubernetes::Helper.new)
+        Adapters::Kubernetes.new(batch)
+      rescue => e
+        puts e.backtrace
+        raise e
       end
-
     end
 
     module Adapters
-      # A class that handles the communication with a resource manager for
-      # submitting/statusing/holding/deleting jobs
-      # @abstract
       class Kubernetes < Adapter
-
-        require 'ood_core/job/adapters/kubernetes/resources'
-        require 'ood_core/job/adapters/kubernetes/helper'
-        require 'json'
 
         using Refinements::ArrayExtensions
         using Refinements::HashExtensions
 
-        attr_reader :config_file, :bin, :restart_polciy, :cluster_name, :mounts
-        attr_reader :using_context, :helper
+        require "ood_core/job/adapters/kubernetes/batch"
 
-        def initialize(options = {})
-          opts = options.to_h.symbolize_keys
+        attr_reader :batch
 
-          @config_file = opts.fetch(:config_file, default_config_file)
-          @bin = opts.fetch(:bin, '/usr/bin/kubectl')
-          @restart_policy = opts.fetch(:restart_policy, 'Never')
-          @cluster_name = opts.fetch(:cluster_name, 'open-ondemand')
-          @mounts = opts.fetch(:mounts, [])
-
-          @using_context = false
-
-          make_kubectl_config(opts)
-
-          @helper = Kubernetes::Helper.new
-        end
-
-        def resource_file(resource_type = 'pod')
-          File.dirname(__FILE__) + "/kubernetes/templates/#{resource_type}.yml.erb"
+        def initialize(batch)
+          @batch = batch
         end
 
         # Submit a job with the attributes defined in the job template instance
@@ -71,16 +55,11 @@ module OodCore
         def submit(script, after: [], afterok: [], afternotok: [], afterany: [])
           raise ArgumentError, 'Must specify the script' if script.nil?
 
-          resource_yml, id = generate_id_yml(script.native)
-          call("#{formatted_ns_cmd} create -f -", stdin: resource_yml)
-
-          id
+          batch.submit(script)
+        rescue Batch::Error => e
+          raise JobAdapterError, e.message
         end
 
-        def generate_id(name)
-          # 2_821_109_907_456 = 36**8
-          name.downcase.tr(' ', '-') + '-' + rand(2_821_109_907_456).to_s(36)
-        end
 
         # Retrieve info for all jobs from the resource manager
         # @abstract Subclass is expected to implement {#info_all}
@@ -94,9 +73,9 @@ module OodCore
         #   adapters can get by without populating the entire Info object
         # @return [Array<Info>] information describing submitted jobs
         def info_all(attrs: nil)
-          cmd = "#{base_cmd} get pods -o json --all-namespaces"
-          output = call(cmd)
-          all_pods_to_info(output)
+          batch.info_all(id.to_s)
+        rescue Batch::Error => e
+          raise JobAdapterError, e.message
         end
 
         # Retrieve info for all jobs for a given owner or owners from the
@@ -169,18 +148,9 @@ module OodCore
         # @param id [#to_s] the id of the job
         # @return [Info] information describing submitted job
         def info(id)
-          pod_json = call_json_output('get', 'pod', id)
-
-          begin
-            service_json = call_json_output('get', 'service', service_name(id))
-            secret_json = call_json_output('get', 'secret', secret_name(id))
-          rescue
-            # it's ok if these don't exist
-            service_json ||= nil
-            secret_json ||= nil
-          end
-
-          helper.info_from_json(pod_json: pod_json, service_json: service_json, secret_json: secret_json)
+          batch.info(id.to_s)
+        rescue Batch::Error => e
+          raise JobAdapterError, e.message
         end
 
         # Retrieve job status from resource manager
@@ -216,224 +186,11 @@ module OodCore
         # @param id [#to_s] the id of the job
         # @return [void]
         def delete(id)
-          call("#{namespaced_cmd} delete pod #{id}")
-
-          begin
-            call("#{namespaced_cmd} delete service #{service_name(id)}")
-            call("#{namespaced_cmd} delete secret #{secret_name(id)}")
-            call("#{namespaced_cmd} delete configmap #{configmap_name(id)}")
-          rescue
-            # FIXME: retries? delete if exists?
-            # just eat the results of deleting services and secrets
-          end
-        end
-
-        def configmap_mount_path
-          '/ood'
-        end
-
-        private
-
-        def username
-          @username ||= Etc.getlogin
-        end
-
-        def run_as_user
-          Etc.getpwnam(username).uid
-        end
-
-        def run_as_group
-          Etc.getpwnam(username).gid
-        end
-
-        def fs_group
-          run_as_group
-        end
-
-        # helper to template resource yml you're going to submit and
-        # create an id.
-        def generate_id_yml(native_data)
-          container = helper.container_from_native(native_data[:container])
-          id = generate_id(container.name)
-          configmap = helper.configmap_from_native(native_data, id)
-          init_containers = helper.init_ctrs_from_native(native_data[:init_ctrs]) if native_data.key?(:init_ctrs)
-          spec = Resources::PodSpec.new(container, init_containers: init_containers)
-
-          template = ERB.new(File.read(resource_file))
-
-          [template.result(binding), id]
-        end
-
-        # helper to call kubectl and get json data back.
-        # verb, resrouce and id are the kubernetes parlance terms.
-        # example: 'kubectl get pod my-pod-id' is verb=get, resource=pod
-        # and  id=my-pod-id
-        def call_json_output(verb, resource, id, stdin: nil)
-          cmd = "#{formatted_ns_cmd} #{verb} #{resource} #{id}"
-          data = call(cmd, stdin: stdin)
-          data = data.empty? ? '{}' : data
-          json_data = JSON.parse(data, symbolize_names: true)
-
-          json_data
-        end
-
-        def service_name(id)
-          helper.service_name(id)
-        end
-
-        def secret_name(id)
-          helper.secret_name(id)
-        end
-
-        def configmap_name(id)
-          helper.configmap_name(id)
-        end
-
-        def namespace
-          default_namespace
-        end
-
-        def default_namespace
-          username
-        end
-
-        def context
-          cluster_name
-        end
-
-        def default_config_file
-          (ENV['KUBECONFIG'] || "#{Dir.home}/.kube/config")
-        end
-
-        def default_auth
-          {
-            type: 'managaged'
-          }.symbolize_keys
-        end
-
-        def default_server
-          {
-            endpoint: 'https://localhost:8080',
-            cert_authority_file: nil
-          }.symbolize_keys
-        end
-
-        def formatted_ns_cmd
-          "#{namespaced_cmd} -o json"
-        end
-
-        def namespaced_cmd
-          "#{base_cmd} --namespace=#{namespace}"
-        end
-
-        def base_cmd
-          base = "#{bin} --kubeconfig=#{config_file}"
-          base << " --context=#{context}" if using_context
-          base
-        end
-
-        def all_pods_to_info(data)
-          json_data = JSON.parse(data, symbolize_names: true)
-          pods = json_data.dig(:items)
-
-          info_array = []
-          pods.each do |pod|
-            info = pod_info_from_json(pod)
-            info_array.push(info) unless info.nil?
-          end
-
-          info_array
-        end
-
-        def pod_info_from_json(pod)
-          hash = helper.pod_info_from_json(pod)
-          Info.new(hash)
-        rescue Kubernetes::Helper::K8sDataError
-          # FIXME: silently eating error, could probably use a logger
-          nil
-        end
-
-        def make_kubectl_config(config)
-          set_cluster(config.fetch(:server, default_server).to_h.symbolize_keys)
-          configure_auth(config.fetch(:auth, default_auth).to_h.symbolize_keys)
-        end
-
-        def configure_auth(auth)
-          type = auth.fetch(:type)
-          return if managed?(type)
-
-          case type
-          when 'gke'
-            set_gke_config(auth)
-          when 'oidc'
-            set_context
-          end
-        end
-
-        def use_context
-          @using_context = true
-        end
-
-        def managed?(type)
-          if type.nil?
-            true # maybe should be false?
-          else
-            type.to_s == 'managed'
-          end
-        end
-
-        def set_gke_config(auth)
-          cred_file = auth.fetch(:svc_acct_file)
-
-          cmd = "gcloud auth activate-service-account --key-file=#{cred_file}"
-          call(cmd)
-
-          set_gke_credentials(auth)
-        end
-
-        def set_gke_credentials(auth)
-
-          zone = auth.fetch(:zone, nil)
-          region = auth.fetch(:region, nil)
-
-          locale = ''
-          locale = "--zone=#{zone}" unless zone.nil?
-          locale = "--region=#{region}" unless region.nil?
-
-          # gke cluster name can probably can differ from what ood calls the cluster
-          cmd = "gcloud container clusters get-credentials #{locale} #{cluster_name}"
-          env = { 'KUBECONFIG' => config_file }
-          call(cmd, env)
-        end
-
-        def set_context
-          cmd = "#{base_cmd} config set-context #{cluster_name}"
-          cmd << " --cluster=#{cluster_name} --namespace=#{namespace}"
-          cmd << " --user=#{username}"
-
-          call(cmd)
-          use_context
-        end
-
-        def set_cluster(config)
-          server = config.fetch(:endpoint)
-          cert = config.fetch(:cert_authority_file, nil)
-
-          cmd = "#{base_cmd} config set-cluster #{cluster_name}"
-          cmd << " --server=#{server}"
-          cmd << " --certificate-authority=#{cert}" unless cert.nil?
-
-          call(cmd)
-        end
-
-        def call(cmd = '', env: {}, stdin: nil)
-          o, error, s = Open3.capture3(env, cmd, stdin_data: stdin.to_s)
-          raise error unless s.success?
-
-          o
+          batch.delete(id.to_s)
+        rescue Batch::Error => e
+          raise JobAdapterError, e.message
         end
       end
-      # end kubernetes class (above)
     end
   end
 end
